+++ GEO450_GRASS_S1/support_functions.py
@@ -1,5 +1,6 @@
 import os
 from GEO450_GRASS_S1.user_data import *
+from grass.pygrass.modules import Module
 
 
 def extract_files_to_list(path_to_folder, datatype):
@@ -70,3 +71,58 @@ def subset_processed_data():
         functions_out_array(outname=subset_path + processed_file_list[k][tmp1 + 5:tmp2 - 4] + "_subset.tif",
                             arr=out_image1,
                             input_file=processed_file_list[k], dtype=np.float32, ras_meta1=ras_meta1)
+
+
+def t_list_output(overwrite_bool):
+    """
+    function to create list of space time raster datasets (SRTDS) created by create_stc function, can be generally used
+    to list SRTDS in GRASS environment
+    :param overwrite_bool: boolean
+        Option of True or False, but True is strongly recommended!
+    """
+    t_list = Module("t.list")
+    t_list(overwrite=overwrite_bool,
+           type="strds",
+           temporaltype="absolute",
+           order="id",
+           columns="id",
+           separator="pipe",
+           output=os.path.join(Paths.main_path, "t_list_output"))
+
+
+def g_list_output(overwrite_bool):
+    """
+    function to create list of raster data created by t_rast_algebra function, can be generally used to list raster data
+    in GRASS environment
+    :param overwrite_bool: boolean
+        Option of True or False, but True is strongly recommended!
+    """
+    g_list = Module("g.list")
+    g_list(overwrite=overwrite_bool,
+           type="raster",
+           separator=",",
+           output=os.path.join(Paths.main_path, "g_list_output"))
+
+
+def t_remove(strds_name):
+    """
+    function to remove space time raster datasets (SRTDS) created by create_stc function, can be generally used to
+    delete active SRTDS in GRASS environment
+    :param strds_name:
+    """
+    strds_remove = Module("t.remove")
+    strds_remove(flags='rf',
+                 type="strds",
+                 inputs=strds_name)
+
+
+def g_remove(raster_name):
+    """
+    function to remove raster products created by t_rast_algebra function, can be generally used to delete active
+    raster data in GRASS environment
+    :param raster_name:
+    """
+    raster_remove = Module("g.remove")
+    raster_remove(flags='fb',
+                  type="raster",
+                  name=raster_name + "@PERMANENT")
